Move useCommentManaget into comments container
